requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==4.2
django-allauth==0.63.4
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-import-export==4.1.1
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
drf-nested-routers==0.94.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
redis==5.0.8
rq==1.16.2
sqlparse==0.5.0
tablib==3.5.0
typing-extensions==4.12.2
tzdata==2024.1
